Adicionados testes de integração total no ambiente de dev usando Jmeter

+++ src/main/java/com/example/openmarket/controller/dto/response/SellerProfileDto.java
@@ -1,5 +1,6 @@
 package com.example.openmarket.controller.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -10,12 +11,12 @@ public class SellerProfileDto {
     private final String storeName;
     private final String storeDescription;
     private final String status;
-    private final Double rating;
+    private final BigDecimal rating;
     private final LocalDateTime createdAt;
 
     public SellerProfileDto(UUID id, String email, String name,
                            String storeName, String storeDescription,
-                           String status, Double rating, LocalDateTime createdAt) {
+                           String status, BigDecimal rating, LocalDateTime createdAt) {
         this.id = id;
         this.email = email;
         this.name = name;
@@ -32,6 +33,6 @@ public class SellerProfileDto {
     public String getStoreName() { return storeName; }
     public String getStoreDescription() { return storeDescription; }
     public String getStatus() { return status; }
-    public Double getRating() { return rating; }
+    public BigDecimal getRating() { return rating; }
     public LocalDateTime getCreatedAt() { return createdAt; }
 }
